+++ example/fibonacci/fib_service.dart
@@ -1,7 +1,10 @@
 import 'dart:async';
 import 'dart:core' as dc;
 
+import 'package:cancelation_token/cancelation_token.dart';
+import 'package:logger/web.dart';
 import 'package:squadron/squadron.dart';
+import 'package:using/using.dart';
 
 import 'generated/fib_service.activator.g.dart';