+++ example/misc/data_service.dart
@@ -1,6 +1,9 @@
 import 'dart:async';
 
+import 'package:cancelation_token/cancelation_token.dart';
+import 'package:logger/web.dart';
 import 'package:squadron/squadron.dart';
+import 'package:using/using.dart';
 
 // ignore: unused_import
 import 'data.dart';