+++ example/perf/echo_service.dart
@@ -4,6 +4,7 @@ import 'dart:convert';
 import 'package:cancelation_token/cancelation_token.dart';
 import 'package:logger/logger.dart';
 import 'package:squadron/squadron.dart';
+import 'package:using/using.dart';
 
 import 'generated/echo_service.activator.g.dart';
 import 'marshalers.dart';