+++ example/test/test_services.dart
@@ -4,7 +4,10 @@
 import 'dart:async';
 import 'dart:typed_data' as typed_data;
 
+import 'package:cancelation_token/cancelation_token.dart';
+import 'package:logger/web.dart';
 import 'package:squadron/squadron.dart' as sq;
+import 'package:using/using.dart';
 
 import 'bigint_marshaler.dart' as marshaler;
 import 'generated/test_services.activator.g.dart';